requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
google-generativeai>=0.3.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
pypdf2==3.0.1
pdfplumber==0.10.3
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
openpyxl==3.1.2
xlrd==2.0.1
flask==3.0.0
flask-cors==4.0.0
reportlab==4.0.7
groq